services/user-service/requirements.txt
email-validator
fastapi
passlib
psycopg2-binary
pydantic
python-dotenv
python-jose
sqlalchemy
uvicorn
pytest
pytest-cov
httpx